update comments and rename NodeResourceUtil to NodeResourceConfig
